Code Refactoring and Incorporating Review Comments
