+++ go/vt/vtgate/vindexes/lookup_internal.go
@@ -1,116 +1,31 @@
 package vindexes
 
-import (
-	"encoding/json"
-	"fmt"
-)
+import "fmt"
 
-func init() {
-	Register("lookup", NewLookup)
-	Register("lookup_unique", NewLookupUnique)
+// lookup implements the functions for the Lookup vindexes.
+type lookup struct {
+	Table              string `json:"table"`
+	From               string `json:"from"`
+	To                 string `json:"to"`
+	sel, ver, ins, del string
 }
 
-// LookupNonUnique defines a vindex that uses a lookup table and create a mapping between id and KeyspaceId.
-//It's NonUnique and a Lookup.
-type LookupNonUnique struct {
-	name string
-	lkp  lookup
+func (lkp *lookup) Init(m map[string]string) {
+	t := m["table"]
+	from := m["from"]
+	to := m["to"]
+
+	lkp.Table = t
+	lkp.From = from
+	lkp.To = to
+	lkp.sel = fmt.Sprintf("select %s from %s where %s = :%s", to, t, from, from)
+	lkp.ver = fmt.Sprintf("select %s from %s where %s = :%s and %s = :%s", from, t, from, from, to, to)
+	lkp.ins = fmt.Sprintf("insert into %s(%s, %s) values(:%s, :%s)", t, from, to, from, to)
+	lkp.del = fmt.Sprintf("delete from %s where %s = :%s and %s = :%s", t, from, from, to, to)
 }
 
-// NewLookup creates a LookupNonUnique vindex.
-func NewLookup(name string, m map[string]string) (Vindex, error) {
-	lookup := &LookupNonUnique{name: name}
-	lookup.lkp.Init(m)
-	return lookup, nil
-}
-
-// String returns the name of the vindex.
-func (vindex *LookupNonUnique) String() string {
-	return vindex.name
-}
-
-// Cost returns the cost of this vindex as 20.
-func (vindex *LookupNonUnique) Cost() int {
-	return 20
-}
-
-// Map returns the corresponding KeyspaceId values for the given ids.
-func (vindex *LookupNonUnique) Map(vcursor VCursor, ids []interface{}) ([][][]byte, error) {
-	return vindex.lkp.MapLookup(vcursor, ids)
-}
-
-// Verify returns true if id maps to ksid.
-func (vindex *LookupNonUnique) Verify(vcursor VCursor, id interface{}, ksid []byte) (bool, error) {
-	return vindex.lkp.VerifyLookup(vcursor, id, ksid)
-}
-
-// Create reserves the id by inserting it into the vindex table.
-func (vindex *LookupNonUnique) Create(vcursor VCursor, id interface{}, ksid []byte) error {
-	return vindex.lkp.CreateLookup(vcursor, id, ksid)
-}
-
-// Delete deletes the entry from the vindex table.
-func (vindex *LookupNonUnique) Delete(vcursor VCursor, ids []interface{}, ksid []byte) error {
-	return vindex.lkp.DeleteLookup(vcursor, ids, ksid)
-}
-
-// MarshalJSON returns a JSON representation of LookupHash.
-func (vindex *LookupNonUnique) MarshalJSON() ([]byte, error) {
-	return json.Marshal(vindex.lkp)
-}
-
-// LookupUnique defines a vindex that uses a lookup table.
-// The table is expected to define the id column as unique. It's
-// Unique and a Lookup.
-type LookupUnique struct {
-	name string
-	lkp  lookup
-}
-
-// NewLookupUnique creates a LookupHashUnique vindex.
-func NewLookupUnique(name string, m map[string]string) (Vindex, error) {
-	lu := &LookupUnique{name: name}
-	lu.lkp.Init(m)
-	return lu, nil
-}
-
-// String returns the name of the vindex.
-func (vindex *LookupUnique) String() string {
-	return vindex.name
-}
-
-// Cost returns the cost of this vindex as 10.
-func (vindex *LookupUnique) Cost() int {
-	return 10
-}
-
-// Map returns the corresponding KeyspaceId values for the given ids.
-func (vindex *LookupUnique) Map(vcursor VCursor, ids []interface{}) ([][]byte, error) {
-	return vindex.lkp.MapLookupUnique(vcursor, ids)
-}
-
-// Verify returns true if id maps to ksid.
-func (vindex *LookupUnique) Verify(vcursor VCursor, id interface{}, ksid []byte) (bool, error) {
-	return vindex.lkp.VerifyLookup(vcursor, id, ksid)
-}
-
-// Create reserves the id by inserting it into the vindex table.
-func (vindex *LookupUnique) Create(vcursor VCursor, id interface{}, ksid []byte) error {
-	return vindex.lkp.CreateLookup(vcursor, id, ksid)
-}
-
-// Delete deletes the entry from the vindex table.
-func (vindex *LookupUnique) Delete(vcursor VCursor, ids []interface{}, ksid []byte) error {
-	return vindex.lkp.DeleteLookup(vcursor, ids, ksid)
-}
-
-// MarshalJSON returns a JSON representation of LookupHashUnique.
-func (vindex *LookupUnique) MarshalJSON() ([]byte, error) {
-	return json.Marshal(vindex.lkp)
-}
-
-// MapLookupUnique is for a unique vindex.
-func (lkp *lookup) MapLookupUnique(vcursor VCursor, ids []interface{}) ([][]byte, error) {
+// MapUniqueLookup is for a Unique Vindex.
+func (lkp *lookup) MapUniqueLookup(vcursor VCursor, ids []interface{}, isHashedIndex bool) ([][]byte, error) {
 	out := make([][]byte, 0, len(ids))
 	for _, id := range ids {
 		result, err := vcursor.Execute(lkp.sel, map[string]interface{}{
@@ -126,13 +41,21 @@ func (lkp *lookup) MapLookupUnique(vcursor VCursor, ids []interface{}) ([][]byte
 		if len(result.Rows) != 1 {
 			return nil, fmt.Errorf("lookup.Map: unexpected multiple results from vindex %s: %v", lkp.Table, id)
 		}
-		out = append(out, result.Rows[0][0].Raw())
+		if isHashedIndex {
+			num, err := getNumber(result.Rows[0][0].ToNative())
+			if err != nil {
+				return nil, fmt.Errorf("lookup.Map: %v", err)
+			}
+			out = append(out, vhash(num))
+		} else {
+			out = append(out, result.Rows[0][0].Raw())
+		}
 	}
 	return out, nil
 }
 
-// MapLookup is for a non-unique vindex.
-func (lkp *lookup) MapLookup(vcursor VCursor, ids []interface{}) ([][][]byte, error) {
+// MapNonUniqueLookup is for a Non-Unique Vindex.
+func (lkp *lookup) MapNonUniqueLookup(vcursor VCursor, ids []interface{}, isHashedIndex bool) ([][][]byte, error) {
 	out := make([][][]byte, 0, len(ids))
 	for _, id := range ids {
 		result, err := vcursor.Execute(lkp.sel, map[string]interface{}{
@@ -142,44 +65,39 @@ func (lkp *lookup) MapLookup(vcursor VCursor, ids []interface{}) ([][][]byte, er
 			return nil, fmt.Errorf("lookup.Map: %v", err)
 		}
 		var ksids [][]byte
-		for _, row := range result.Rows {
-			ksids = append(ksids, row[0].Raw())
+		if isHashedIndex {
+			for _, row := range result.Rows {
+				num, err := getNumber(row[0].ToNative())
+				if err != nil {
+					return nil, fmt.Errorf("lookup.Map: %v", err)
+				}
+				ksids = append(ksids, vhash(num))
+			}
+		} else {
+			for _, row := range result.Rows {
+				ksids = append(ksids, row[0].Raw())
+			}
 		}
 		out = append(out, ksids)
 	}
 	return out, nil
 }
 
-// Create creates an association between id and keyspaceid by inserting a row in the vindex table.
-func (lkp *lookup) CreateLookup(vcursor VCursor, id interface{}, ksid []byte) error {
-	if _, err := vcursor.Execute(lkp.ins, map[string]interface{}{
-		lkp.From: id,
-		lkp.To:   ksid,
-	}); err != nil {
-		return fmt.Errorf("lookup.Create: %v", err)
-	}
-	return nil
-}
-
-// Delete deletes the association between ids and keyspaceid.
-func (lkp *lookup) DeleteLookup(vcursor VCursor, ids []interface{}, ksid []byte) error {
-	bindvars := map[string]interface{}{
-		lkp.To: ksid,
-	}
-	for _, id := range ids {
-		bindvars[lkp.From] = id
-		if _, err := vcursor.Execute(lkp.del, bindvars); err != nil {
-			return fmt.Errorf("lookup.Delete: %v", err)
+// Verify returns true if id maps to ksid.
+func (lkp *lookup) Verify(vcursor VCursor, id interface{}, ksid []byte, isHashedIndex bool) (bool, error) {
+	var val interface{}
+	var err error
+	if isHashedIndex {
+		val, err = vunhash(ksid)
+		if err != nil {
+			return false, fmt.Errorf("lookup.Verify: %v", err)
 		}
+	} else {
+		val = ksid
 	}
-	return nil
-}
-
-// VerifyLookup returns true if id maps to ksid.
-func (lkp *lookup) VerifyLookup(vcursor VCursor, id interface{}, ksid []byte) (bool, error) {
 	result, err := vcursor.Execute(lkp.ver, map[string]interface{}{
 		lkp.From: id,
-		lkp.To:   ksid,
+		lkp.To:   val,
 	})
 	if err != nil {
 		return false, fmt.Errorf("lookup.Verify: %v", err)
@@ -188,4 +106,49 @@ func (lkp *lookup) VerifyLookup(vcursor VCursor, id interface{}, ksid []byte) (b
 		return false, nil
 	}
 	return true, nil
+}
+
+// Create creates an association between id and ksid by inserting a row in the vindex table.
+func (lkp *lookup) Create(vcursor VCursor, id interface{}, ksid []byte, isHashedIndex bool) error {
+	var val interface{}
+	var err error
+	if isHashedIndex {
+		val, err = vunhash(ksid)
+		if err != nil {
+			return fmt.Errorf("lookup.Create: %v", err)
+		}
+	} else {
+		val = ksid
+	}
+	if _, err := vcursor.Execute(lkp.ins, map[string]interface{}{
+		lkp.From: id,
+		lkp.To:   val,
+	}); err != nil {
+		return fmt.Errorf("lookup.Create: %v", err)
+	}
+	return nil
+}
+
+// Delete deletes the association between ids and ksid.
+func (lkp *lookup) Delete(vcursor VCursor, ids []interface{}, ksid []byte, isHashedIndex bool) error {
+	var val interface{}
+	var err error
+	if isHashedIndex {
+		val, err = vunhash(ksid)
+		if err != nil {
+			return fmt.Errorf("lookup.Delete: %v", err)
+		}
+	} else {
+		val = ksid
+	}
+	bindvars := map[string]interface{}{
+		lkp.To: val,
+	}
+	for _, id := range ids {
+		bindvars[lkp.From] = id
+		if _, err := vcursor.Execute(lkp.del, bindvars); err != nil {
+			return fmt.Errorf("lookup.Delete: %v", err)
+		}
+	}
+	return nil
 }
