+++ go/vt/vtgate/vindexes/lookup_test.go
